refactor: cleaner responsibilities across files :truck:
